refactor(runtime): share common event handler code between trigger and enricher factories
